Update usability and prepare for tests.
Depending on the use-case, one wants to either run all commands inside the
Docker container using the `web` service or one uses `pipenv` to run stuff
locally. This commit fixes a bug, where usage of `pipenv run` was not possible.

Now there are two ways of running commands:
- Using `docker-compose`, thus inside the Docker container: `docker-compose run
--rm web python manage.py help`
- Using `pipenv`: `pipenv run python manage.py help`.

We now check whether we're running inside of the Docker container (lazily
checking for the `PYTHONBUFFERED` environment variable) and set the database
host accordingly.